[API] Move PickupCart command and handler to cart namespace
